UI is basically done.
But something went haywire as all hell when I was tying everything together directly to get it working as one program. I ended up losing like 3 hours of work that I'll have to redo tomorrow or the next day. But I'll get us a working program by the end of this weekend come Hell or high water.